requirements.txt
streamlit>=1.22.0
SpeechRecognition>=3.10.0
textblob>=0.17.1
openai>=0.27.0
requests>=2.28.0
Pillow>=9.5.0
python-dotenv>=1.0.0 